refactor: change price fields from BigDecimal to Integer with TWD pricing
- Replace all BigDecimal price fields with Integer type across entities, DTOs, and services
- Remove all BigDecimal imports and calculations
- Update database migration to change column types from DECIMAL to INT
- Adjust all product prices to reasonable TWD amounts:
  * Test products: NT$500-800
  * MacBook Pro M4: NT$65,000
  * iPhone 15 Pro: NT$36,900
  * AirPods Pro: NT$7,990
  * iPad Air: NT$19,900
  * Apple Watch Series 9: NT$13,900
  * Magic Keyboard: NT$4,790
  * Magic Mouse: NT$2,590
  * USB-C Cable: NT$590
  * MagSafe Charger: NT$1,290
  * Studio Display: NT$49,900
- Update tracking message format from %.2f to %d for integer display
- Update all test files with appropriate TWD test prices (1000, 2000, 3000, 4000)

+++ backend/src/main/java/com/etrex/oms/controller/TestDataController.java
@@ -14,7 +14,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -84,7 +83,7 @@ public class TestDataController {
     public static class ProductSeedRequest {
         private String name;
         private String description;
-        private BigDecimal price;
+        private Integer price;
         private Integer stock;
         private String status;
     }
