+++ backend/src/main/java/com/etrex/oms/entity/OrderItem.java
@@ -8,8 +8,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Entity
 @Table(name = "order_items")
 @Data
@@ -32,5 +30,5 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(nullable = false)
-    private BigDecimal price;
+    private Integer price;
 }
